Change is_admin to tinyint

+++ config/schema.sql
@@ -6,7 +6,7 @@ CREATE TABLE users(
     email VARCHAR(60) UNIQUE,
     first_name VARCHAR(56),
     last_name VARCHAR(56),
-    is_admin INT(2),
+    is_admin TINYINT(1),
   CONSTRAINT users_pk PRIMARY KEY (id)
 );
